moves utility function
