+++ corehq/ex-submodules/dimagi/utils/django/email.py
@@ -1,11 +1,7 @@
 from abc import ABC, abstractmethod
-from smtplib import SMTPDataError, SMTPSenderRefused
 
 from django.conf import settings
 from django.core.mail import get_connection as django_get_connection
-from django.core.mail.message import EmailMultiAlternatives
-from dimagi.utils.logging import notify_exception
-from django.utils.translation import gettext as _
 
 from corehq.util.metrics import metrics_gauge, metrics_counter
 from corehq.util.metrics.const import MPM_LIVESUM
@@ -78,96 +74,6 @@ def get_valid_recipients(recipients, domain=None):
             'email_domain': email_domain,
         }, multiprocess_mode=MPM_LIVESUM)
     return [recipient for recipient in recipients if recipient not in bounced_emails]
-
-
-def send_HTML_email(subject, recipient, html_content, text_content=None,
-                    cc=None, email_from=settings.DEFAULT_FROM_EMAIL,
-                    file_attachments=None, bcc=None,
-                    smtp_exception_skip_list=None, messaging_event_id=None,
-                    domain=None, use_domain_gateway=False):
-    recipient_list = [_f for _f in recipient if _f] if not isinstance(recipient, str) else [recipient]
-    filtered_recipient_list = get_valid_recipients(recipient_list, domain)
-    bounced_addresses = list(set(recipient_list) - set(filtered_recipient_list))
-    if bounced_addresses and messaging_event_id:
-        mark_local_bounced_email(bounced_addresses, messaging_event_id)
-
-    if not filtered_recipient_list:
-        # todo address root issues by throwing a real error to catch upstream
-        #  fail silently for now to fix time-sensitive SES issue
-        return
-
-    if not isinstance(html_content, str):
-        html_content = html_content.decode('utf-8')
-
-    if not text_content:
-        text_content = getattr(settings, 'NO_HTML_EMAIL_MESSAGE',
-                               NO_HTML_EMAIL_MESSAGE)
-    elif not isinstance(text_content, str):
-        text_content = text_content.decode('utf-8')
-    configuration = get_email_configuration(domain, use_domain_gateway, email_from)
-    headers = {'From': configuration.from_email}  # From-header
-
-    if configuration.return_path_email:
-        headers['Return-Path'] = configuration.return_path_email
-
-    if messaging_event_id is not None:
-        headers[COMMCARE_MESSAGE_ID_HEADER] = messaging_event_id
-    if configuration.SES_configuration_set is not None:
-        headers[SES_CONFIGURATION_SET_HEADER] = configuration.SES_configuration_set
-
-    msg = EmailMultiAlternatives(subject, text_content, configuration.from_email,
-                                 filtered_recipient_list, headers=headers,
-                                 connection=configuration.connection, cc=cc, bcc=bcc)
-    for file in (file_attachments or []):
-        if file:
-            msg.attach(file["title"], file["file_obj"].getvalue(),
-                       file["mimetype"])
-    msg.attach_alternative(html_content, "text/html")
-
-    try:
-        msg.send()
-    except SMTPDataError as e:
-        # If the SES configuration has not been properly set up, resend the message
-        if (
-            "Configuration Set does not exist" in repr(e.smtp_error)
-            and SES_CONFIGURATION_SET_HEADER in msg.extra_headers
-        ):
-            del msg.extra_headers[SES_CONFIGURATION_SET_HEADER]
-            msg.send()
-            notify_exception(None, message="SES Configuration Set missing", details={'error': e})
-        else:
-            raise
-    except SMTPSenderRefused as e:
-
-        if smtp_exception_skip_list and e.smtp_code in smtp_exception_skip_list:
-            raise e
-        else:
-            error_subject = _('ERROR: Could not send "%(subject)s"') % {
-                'subject': subject,
-            }
-
-            if e.smtp_code in LARGE_FILE_SIZE_ERROR_CODES:
-                error_text = _('Could not send email: file size is too large.')
-            else:
-                error_text = e.smtp_error
-            error_text = '%s\n\n%s' % (
-                error_text,
-                _('Please contact %(support_email)s for assistance.') % {
-                    'support_email': settings.SUPPORT_EMAIL,
-                },
-            )
-
-            error_msg = EmailMultiAlternatives(
-                error_subject,
-                error_text,
-                configuration.from_email,
-                filtered_recipient_list,
-                headers=headers,
-                connection=configuration.connection,
-                cc=cc,
-                bcc=bcc,
-            )
-            error_msg.send()
 
 
 def get_email_configuration(domain: str, use_domain_gateway: bool = True,
